README.md
This is me, I am Yes. I'm looking to create a mod that adds an assortment of intelligent alien species into the game Rimworld, each with their own lore and backstory. What's important to me is making something I'm happy with, and making sure that the aliens are something someone has never seen before!